refactor: computer and lab saving were add in their respective repository classes
